Add RequestController.prepare
