+++ lib/src/application/application.dart
@@ -168,7 +168,7 @@ class Application<RequestSinkType extends RequestSink> {
 
     RequestSink sink = sinkMirror.newInstance(new Symbol(""), [config]).reflectee;
     sink.setupRouter(sink.router);
-    sink.router.finalize();
+    sink.router.prepare();
 
     return sink.documentAPI(resolver);
   }
