+++ test/auth/authorizer_test.dart
@@ -415,7 +415,7 @@ void main() {
 Future<HttpServer> enableAuthorizer(Authorizer authorizer) async {
   var router = new Router();
   router.route("/").pipe(authorizer).listen(respond);
-  router.finalize();
+  router.prepare();
 
   var server = await HttpServer.bind(InternetAddress.ANY_IP_V4, 8000);
   server.map((httpReq) => new Request(httpReq)).listen((r) {
