+++ test/base/controller_body_binding_test.dart
@@ -266,7 +266,7 @@ Future<HttpServer> enableController(String pattern, Type controller) async {
   var router = new Router();
   router.route(pattern).generate(
           () => reflectClass(controller).newInstance(new Symbol(""), []).reflectee);
-  router.finalize();
+  router.prepare();
 
   var server = await HttpServer.bind(InternetAddress.ANY_IP_V4, 4040);
   server.map((httpReq) => new Request(httpReq)).listen(router.receive);
